+++ src/main/java/com/github/johnmedlockdev/project1john/Application.java
@@ -1,7 +1,5 @@
 package com.github.johnmedlockdev.project1john;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
@@ -9,11 +7,10 @@ import java.io.IOException;
 
 @SpringBootApplication
 public class Application {
+
     public static void main(String[] args) throws IOException {
         SpringApplication.run(Application.class, args);
     }
-
-        private static final Logger log = LoggerFactory.getLogger(Application.class);
 }