+++ src/main/java/com/github/johnmedlockdev/project1john/controllers/StaticController.java
@@ -1,8 +1,11 @@
 package com.github.johnmedlockdev.project1john.controllers;
 
+import com.github.johnmedlockdev.project1john.models.SparkModel;
+import com.github.johnmedlockdev.project1john.repositories.SparkRepository;
 import org.apache.spark.api.java.JavaPairRDD;
 import org.apache.spark.api.java.JavaRDD;
 import org.apache.spark.api.java.JavaSparkContext;
+import org.apache.spark.storage.StorageLevel;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
 import org.springframework.stereotype.Controller;
@@ -26,6 +29,9 @@ public class StaticController {
     @Autowired
     JavaSparkContext sc;
 
+    @Autowired
+    SparkRepository repository;
+
     private String fileName;
 
     @RequestMapping(value = "/")
@@ -44,14 +50,14 @@ public class StaticController {
         } catch (Exception exe) {
             exe.printStackTrace();
         }
-        return new RedirectView("spark"); // should go to d3 graph
+        return new RedirectView("spark");
     }
 
     @RequestMapping(value = "/spark")
-    @ResponseBody
     public RedirectView spark() {
 
-        JavaRDD<String> allRows = sc.textFile(fileName);
+        JavaRDD<String> allRows = sc.textFile(fileName).persist(StorageLevel.MEMORY_ONLY());
+
         List<String> headers = Arrays.asList(allRows.take(1).get(0).split(","));
 
         System.out.println(headers);
@@ -66,66 +72,20 @@ public class StaticController {
         JavaRDD<Double> column = data.map(
                 x -> Double.valueOf(x.split(",")[headers.
                         indexOf(field)]));
-//        Persist session in memory
-//        column.persist(StorageLevel.MEMORY_ONLY());
 
         JavaRDD<Double> ceil = column.map(Math::ceil);
 
         JavaPairRDD<Double, Integer> priceMap = ceil.mapToPair((f) -> new Tuple2<>(f, 1));
 
-        JavaPairRDD<Double, Integer> countNames = priceMap.reduceByKey((x, y) -> ((int) x + (int) y));
-        System.out.println(countNames.collect());
+        JavaPairRDD<Double, Integer> countValues = priceMap.reduceByKey((x, y) -> ((int) x + (int) y));
 
-
-//
-//        for(Double x : ceil.collect()){
-//            System.out.println(x);
-//        }
-
-
-        return new RedirectView("/");
-
-
+        List<Tuple2<Double, Integer>> list = countValues.collect();
+        for (Tuple2<Double, Integer> x : list) {
+            repository.save(new SparkModel(x._1, x._2));
+        }
+        return new RedirectView("/graph");
     }
 }
-//      ==   how you would use spark over array ==
-//
-//        SparkConf conf = new SparkConf().setAppName("appName").setMaster("local");
-//        JavaSparkContext sc = new JavaSparkContext(conf);
-//
-//        List<Integer> data = Arrays.asList(1, 2, 3, 4, 5);
-//        JavaRDD<Integer> distData = sc.parallelize(data);
-
-//       ===  how you would use spark over file  ==
-//
-//        JavaRDD<String> lines = sc.textFile("data.txt");
-//        JavaRDD<Integer> lineLengths = lines.map(s -> s.length());
-//        int totalLength = lineLengths.reduce((a, b) -> a + b);
-//
-//        == how you would save data ==
-//
-//        lineLengths.persist(StorageLevel.MEMORY_ONLY());
-//
-//        JavaRDD<String> lines = sc.textFile("data.txt");
-//        JavaPairRDD<String, Integer> pairs = lines.mapToPair(s -> new Tuple2(s, 1));
-//        JavaPairRDD<String, Integer> counts = pairs.reduceByKey((a, b) -> a + b);
-
-//        Some notes on reading files with Spark:
-//	*
-//        If using a path on the local filesystem,
-//        the file must also be accessible at the same path on worker nodes.
-//        Either copy the file to all workers or use a network-mounted shared file system.
-//                *
-//                All of Spark’s file-based input methods,
-//                including textFile, support running on directories, compressed files, and wildcards as well.
-//                For example, you can use textFile("/my/directory"), textFile("/my/directory/*.txt"),
-//                and textFile("/my/directory/*.gz").
-
-//                Apart from text files, Spark’s Java API also supports several other data formats:
-//	*
-//        JavaSparkContext.wholeTextFiles lets you read a directory containing multiple small text files,
-//        and returns each of them as (filename, content) pairs. This is in contrast with textFile,
-//        which would return one record per line in each file.
 
 
 // // TODO: 5/9/2020 add a buffer to multipart